third_party/WebKit/LayoutTests/inspector/open-with-rendering-option-enabled.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script>

function test()
{
    WebInspector.settings.showRenderingViewInDrawer.set(true);
    InspectorTest.dumpLoadedModules();
    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>This test checks load with rendering options turned on by default.</p>
</body>
</html>
