third_party/WebKit/LayoutTests/inspector/open-with-rendering-option-enabled-expected.txt

This test checks load with rendering options turned on by default.

Loaded modules:
    common
    host
    sdk
    ui
    components
    bindings
    screencast
    toolbox
    main
    elements
